+++ lib/gui/app/components/source-selector/source-selector.tsx
@@ -14,15 +14,23 @@
  * limitations under the License.
  */
 
-import { faFile, faLink } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import FileSvg from '@fortawesome/fontawesome-free/svgs/solid/file.svg';
+import LinkSvg from '@fortawesome/fontawesome-free/svgs/solid/link.svg';
+import ExclamationTriangleSvg from '@fortawesome/fontawesome-free/svgs/solid/exclamation-triangle.svg';
 import { sourceDestination } from 'etcher-sdk';
 import { ipcRenderer, IpcRendererEvent } from 'electron';
 import * as _ from 'lodash';
 import { GPTPartition, MBRPartition } from 'partitioninfo';
 import * as path from 'path';
 import * as React from 'react';
-import { ButtonProps, Card as BaseCard, Input, Modal, Txt } from 'rendition';
+import {
+	Flex,
+	ButtonProps,
+	Modal as SmallModal,
+	Txt,
+	Card as BaseCard,
+	Input,
+} from 'rendition';
 import styled from 'styled-components';
 
 import * as errors from '../../../../shared/errors';
@@ -38,8 +46,10 @@ import { replaceWindowsNetworkDriveLetter } from '../../os/windows-network-drive
 import {
 	ChangeButton,
 	DetailsText,
+	Modal,
 	StepButton,
 	StepNameButton,
+	ScrollableFlex,
 } from '../../styled-components';
 import { colors } from '../../theme';
 import { middleEllipsis } from '../../utils/middle-ellipsis';
@@ -49,19 +59,24 @@ import ImageSvg from '../../../assets/image.svg';
 
 const recentUrlImagesKey = 'recentUrlImages';
 
-function normalizeRecentUrlImages(urls: any): string[] {
+function normalizeRecentUrlImages(urls: any[]): URL[] {
 	if (!Array.isArray(urls)) {
 		urls = [];
 	}
-	return _.chain(urls)
-		.filter(_.isString)
-		.reject(_.isEmpty)
-		.uniq()
-		.takeRight(5)
-		.value();
+	urls = urls
+		.map((url) => {
+			try {
+				return new URL(url);
+			} catch (error) {
+				// Invalid URL, skip
+			}
+		})
+		.filter((url) => url !== undefined);
+	urls = _.uniqBy(urls, (url) => url.href);
+	return urls.slice(urls.length - 5);
 }
 
-function getRecentUrlImages(): string[] {
+function getRecentUrlImages(): URL[] {
 	let urls = [];
 	try {
 		urls = JSON.parse(localStorage.getItem(recentUrlImagesKey) || '[]');
@@ -71,11 +86,9 @@ function getRecentUrlImages(): string[] {
 	return normalizeRecentUrlImages(urls);
 }
 
-function setRecentUrlImages(urls: string[]) {
-	localStorage.setItem(
-		recentUrlImagesKey,
-		JSON.stringify(normalizeRecentUrlImages(urls)),
-	);
+function setRecentUrlImages(urls: URL[]) {
+	const normalized = normalizeRecentUrlImages(urls.map((url: URL) => url.href));
+	localStorage.setItem(recentUrlImagesKey, JSON.stringify(normalized));
 }
 
 const Card = styled(BaseCard)`
@@ -103,37 +116,42 @@ function getState() {
 	};
 }
 
-const URLSelector = ({ done }: { done: (imageURL: string) => void }) => {
+const URLSelector = ({
+	done,
+	cancel,
+}: {
+	done: (imageURL: string) => void;
+	cancel: () => void;
+}) => {
 	const [imageURL, setImageURL] = React.useState('');
-	const [recentImages, setRecentImages]: [
-		string[],
-		(value: React.SetStateAction<string[]>) => void,
-	] = React.useState([]);
+	const [recentImages, setRecentImages] = React.useState<URL[]>([]);
 	const [loading, setLoading] = React.useState(false);
 	React.useEffect(() => {
 		const fetchRecentUrlImages = async () => {
-			const recentUrlImages: string[] = await getRecentUrlImages();
+			const recentUrlImages: URL[] = await getRecentUrlImages();
 			setRecentImages(recentUrlImages);
 		};
 		fetchRecentUrlImages();
 	}, []);
 	return (
 		<Modal
+			cancel={cancel}
 			primaryButtonProps={{
-				disabled: loading,
+				className: loading || !imageURL ? 'disabled' : '',
 			}}
 			done={async () => {
 				setLoading(true);
-				const sanitizedRecentUrls = normalizeRecentUrlImages([
-					...recentImages,
+				const urlStrings = recentImages.map((url: URL) => url.href);
+				const normalizedRecentUrls = normalizeRecentUrlImages([
+					...urlStrings,
 					imageURL,
 				]);
-				setRecentUrlImages(sanitizedRecentUrls);
+				setRecentUrlImages(normalizedRecentUrls);
 				await done(imageURL);
 			}}
 		>
-			<label style={{ width: '100%' }}>
-				<Txt mb="10px" fontSize="20px">
+			<Flex style={{ width: '100%' }} flexDirection="column">
+				<Txt mb="10px" fontSize="24px">
 					Use Image URL
 				</Txt>
 				<Input
@@ -144,26 +162,31 @@ const URLSelector = ({ done }: { done: (imageURL: string) => void }) => {
 						setImageURL(evt.target.value)
 					}
 				/>
-			</label>
-			{!_.isEmpty(recentImages) && (
-				<div>
-					Recent
-					<Card
-						style={{ padding: '10px 15px' }}
-						rows={_.map(recentImages, (recent) => (
-							<Txt
-								key={recent}
-								onClick={() => {
-									setImageURL(recent);
-								}}
-							>
-								<span>
-									{_.last(_.split(recent, '/'))} - {recent}
-								</span>
-							</Txt>
-						))}
-					/>
-				</div>
+			</Flex>
+			{recentImages.length > 0 && (
+				<Flex flexDirection="column" height="78.6%">
+					<Txt fontSize={18}>Recent</Txt>
+					<ScrollableFlex flexDirection="column">
+						<Card
+							p="10px 15px"
+							rows={recentImages
+								.map((recent) => (
+									<Txt
+										key={recent.href}
+										onClick={() => {
+											setImageURL(recent.href);
+										}}
+										style={{
+											overflowWrap: 'break-word',
+										}}
+									>
+										{recent.pathname.split('/').pop()} - {recent.href}
+									</Txt>
+								))
+								.reverse()}
+						/>
+					</ScrollableFlex>
+				</Flex>
 			)}
 		</Modal>
 	);
@@ -225,7 +248,7 @@ export class SourceSelector extends React.Component<
 	SourceSelectorProps,
 	SourceSelectorState
 > {
-	private unsubscribe: () => void;
+	private unsubscribe: (() => void) | undefined;
 	private afterSelected: SourceSelectorProps['afterSelected'];
 
 	constructor(props: SourceSelectorProps) {
@@ -255,13 +278,13 @@ export class SourceSelector extends React.Component<
 	}
 
 	public componentWillUnmount() {
-		this.unsubscribe();
+		this.unsubscribe?.();
 		ipcRenderer.removeListener('select-image', this.onSelectImage);
 	}
 
 	private async onSelectImage(_event: IpcRendererEvent, imagePath: string) {
 		const isURL =
-			_.startsWith(imagePath, 'https://') || _.startsWith(imagePath, 'http://');
+			imagePath.startsWith('https://') || imagePath.startsWith('http://');
 		await this.selectImageByPath({
 			imagePath,
 			SourceType: isURL ? sourceDestination.Http : sourceDestination.File,
@@ -335,8 +358,8 @@ export class SourceSelector extends React.Component<
 			});
 		} else {
 			if (
-				!_.startsWith(imagePath, 'https://') &&
-				!_.startsWith(imagePath, 'http://')
+				!imagePath.startsWith('https://') &&
+				!imagePath.startsWith('http://')
 			) {
 				const invalidImageError = errors.createUserError({
 					title: 'Unsupported protocol',
@@ -464,70 +487,64 @@ export class SourceSelector extends React.Component<
 
 		return (
 			<>
-				<div
-					className="box text-center relative"
+				<Flex
+					flexDirection="column"
+					alignItems="center"
 					onDrop={this.onDrop}
 					onDragEnter={this.onDragEnter}
 					onDragOver={this.onDragOver}
 				>
-					<div className="center-block">
-						<SVGIcon
-							contents={imageLogo}
-							fallback={<ImageSvg width="40px" height="40px" />}
-						/>
-					</div>
+					<SVGIcon
+						contents={imageLogo}
+						fallback={ImageSvg}
+						style={{
+							marginBottom: 30,
+						}}
+					/>
 
-					<div className="space-vertical-large">
-						{hasImage ? (
-							<>
-								<StepNameButton
-									plain
-									fontSize={16}
-									onClick={this.showSelectedImageDetails}
-									tooltip={imageName || imageBasename}
-								>
-									{middleEllipsis(imageName || imageBasename, 20)}
-								</StepNameButton>
-								{!flashing && (
-									<ChangeButton plain mb={14} onClick={this.reselectImage}>
-										Remove
-									</ChangeButton>
-								)}
-								<DetailsText>
-									{shared.bytesToClosestUnit(imageSize)}
-								</DetailsText>
-							</>
-						) : (
-							<>
-								<FlowSelector
-									key="Flash from file"
-									flow={{
-										onClick: this.openImageSelector,
-										label: 'Flash from file',
-										icon: <FontAwesomeIcon icon={faFile} />,
-									}}
-								/>
-								<FlowSelector
-									key="Flash from URL"
-									flow={{
-										onClick: this.openURLSelector,
-										label: 'Flash from URL',
-										icon: <FontAwesomeIcon icon={faLink} />,
-									}}
-								/>
-							</>
-						)}
-					</div>
-				</div>
+					{hasImage ? (
+						<>
+							<StepNameButton
+								plain
+								onClick={this.showSelectedImageDetails}
+								tooltip={imageName || imageBasename}
+							>
+								{middleEllipsis(imageName || imageBasename, 20)}
+							</StepNameButton>
+							{!flashing && (
+								<ChangeButton plain mb={14} onClick={this.reselectImage}>
+									Remove
+								</ChangeButton>
+							)}
+							<DetailsText>{shared.bytesToClosestUnit(imageSize)}</DetailsText>
+						</>
+					) : (
+						<>
+							<FlowSelector
+								key="Flash from file"
+								flow={{
+									onClick: this.openImageSelector,
+									label: 'Flash from file',
+									icon: <FileSvg height="1em" fill="currentColor" />,
+								}}
+							/>
+							<FlowSelector
+								key="Flash from URL"
+								flow={{
+									onClick: this.openURLSelector,
+									label: 'Flash from URL',
+									icon: <LinkSvg height="1em" fill="currentColor" />,
+								}}
+							/>
+						</>
+					)}
+				</Flex>
 
 				{this.state.warning != null && (
-					<Modal
+					<SmallModal
 						titleElement={
 							<span>
-								<span
-									style={{ color: '#d9534f' }}
-									className="glyphicon glyphicon-exclamation-sign"
-								></span>{' '}
+								<ExclamationTriangleSvg fill="#fca321" height="1em" />{' '}
 								<span>{this.state.warning.title}</span>
 							</span>
 						}
@@ -544,11 +561,11 @@ export class SourceSelector extends React.Component<
 						<ModalText
 							dangerouslySetInnerHTML={{ __html: this.state.warning.message }}
 						/>
-					</Modal>
+					</SmallModal>
 				)}
 
 				{showImageDetails && (
-					<Modal
+					<SmallModal
 						title="Image"
 						done={() => {
 							this.setState({ showImageDetails: false });
@@ -562,11 +579,16 @@ export class SourceSelector extends React.Component<
 							<Txt.span bold>Path: </Txt.span>
 							<Txt.span>{imagePath}</Txt.span>
 						</Txt.p>
-					</Modal>
+					</SmallModal>
 				)}
 
 				{showURLSelector && (
 					<URLSelector
+						cancel={() => {
+							this.setState({
+								showURLSelector: false,
+							});
+						}}
 						done={async (imageURL: string) => {
 							// Avoid analytics and selection state changes
 							// if no file was resolved from the dialog.