+++ lib/gui/app/components/target-selector/target-selector-button.tsx
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
+import ExclamationTriangleSvg from '@fortawesome/fontawesome-free/svgs/solid/exclamation-triangle.svg';
 import { Drive as DrivelistDrive } from 'drivelist';
-import * as _ from 'lodash';
 import * as React from 'react';
-import { Txt } from 'rendition';
-import { default as styled } from 'styled-components';
+import { Flex, FlexProps, Txt } from 'rendition';
 
 import {
 	getDriveImageCompatibilityStatuses,
@@ -34,10 +33,6 @@ import {
 } from '../../styled-components';
 import { middleEllipsis } from '../../utils/middle-ellipsis';
 
-const TargetDetail = styled((props) => <Txt.span {...props}></Txt.span>)`
-	float: ${({ float }) => float};
-`;
-
 interface TargetSelectorProps {
 	targets: any[];
 	disabled: boolean;
@@ -49,24 +44,26 @@ interface TargetSelectorProps {
 	image: Image;
 }
 
-function DriveCompatibilityWarning(props: {
+function DriveCompatibilityWarning({
+	drive,
+	image,
+	...props
+}: {
 	drive: DrivelistDrive;
 	image: Image;
-}) {
+} & FlexProps) {
 	const compatibilityWarnings = getDriveImageCompatibilityStatuses(
-		props.drive,
-		props.image,
+		drive,
+		image,
 	);
 	if (compatibilityWarnings.length === 0) {
 		return null;
 	}
-	const messages = _.map(compatibilityWarnings, 'message');
+	const messages = compatibilityWarnings.map((warning) => warning.message);
 	return (
-		<Txt.span
-			className="glyphicon glyphicon-exclamation-sign"
-			ml={2}
-			tooltip={messages.join(', ')}
-		/>
+		<Flex tooltip={messages.join(', ')} {...props}>
+			<ExclamationTriangleSvg fill="currentColor" height="1em" />
+		</Flex>
 	);
 }
 
@@ -86,7 +83,11 @@ export function TargetSelector(props: TargetSelectorProps) {
 					</ChangeButton>
 				)}
 				<DetailsText>
-					<DriveCompatibilityWarning drive={target} image={props.image} />
+					<DriveCompatibilityWarning
+						drive={target}
+						image={props.image}
+						mr={2}
+					/>
 					{bytesToClosestUnit(target.size)}
 				</DetailsText>
 			</>
@@ -104,21 +105,19 @@ export function TargetSelector(props: TargetSelectorProps) {
 					} ${bytesToClosestUnit(target.size)}`}
 					px={21}
 				>
-					<Txt.span>
-						<DriveCompatibilityWarning drive={target} image={props.image} />
-						<TargetDetail float="left">
-							{middleEllipsis(target.description, 14)}
-						</TargetDetail>
-						<TargetDetail float="right">
-							{bytesToClosestUnit(target.size)}
-						</TargetDetail>
-					</Txt.span>
+					<DriveCompatibilityWarning
+						drive={target}
+						image={props.image}
+						mr={2}
+					/>
+					<Txt mr={2}>{middleEllipsis(target.description, 14)}</Txt>
+					<Txt>{bytesToClosestUnit(target.size)}</Txt>
 				</DetailsText>,
 			);
 		}
 		return (
 			<>
-				<StepNameButton plain tooltip={props.tooltip} fontSize={16}>
+				<StepNameButton plain tooltip={props.tooltip}>
 					{targets.length} Targets
 				</StepNameButton>
 				{!props.flashing && (