+++ lib/gui/app/modules/image-writer.ts
@@ -93,7 +93,11 @@ function terminateServer() {
 }
 
 function writerArgv(): string[] {
-	let entryPoint = electron.remote.app.getAppPath();
+	let entryPoint = path.join(
+		electron.remote.app.getAppPath(),
+		'generated',
+		'child-writer.js',
+	);
 	// AppImages run over FUSE, so the files inside the mount point
 	// can only be accessed by the user that mounted the AppImage.
 	// This means we can't re-spawn Etcher as root from the same
@@ -132,16 +136,13 @@ interface FlashResults {
 
 /**
  * @summary Perform write operation
- *
- * @description
- * This function is extracted for testing purposes.
  */
-export async function performWrite(
+async function performWrite(
 	image: string,
 	drives: DrivelistDrive[],
 	onProgress: sdk.multiWrite.OnProgressFunction,
 	source: SourceOptions,
-): Promise<{ cancelled?: boolean }> {
+): Promise<FlashResults> {
 	let cancelled = false;
 	ipc.serve();
 	const {
@@ -240,7 +241,6 @@ export async function performWrite(
 						title: 'The writer process ended unexpectedly',
 						description:
 							'Please try again, and contact the Etcher team if the problem persists',
-						code: 'ECHILDDIED',
 					}),
 				);
 				return;
@@ -261,6 +261,8 @@ export async function flash(
 	image: string,
 	drives: DrivelistDrive[],
 	source: SourceOptions,
+	// This function is a parameter so it can be mocked in tests
+	write = performWrite,
 ): Promise<void> {
 	if (flashState.isFlashing()) {
 		throw new Error('There is already a flash in progress');
@@ -285,8 +287,7 @@ export async function flash(
 	analytics.logEvent('Flash', analyticsData);
 
 	try {
-		// Using it from exports so it can be mocked during tests
-		const result = await exports.performWrite(
+		const result = await write(
 			image,
 			drives,
 			flashState.setProgressState,